+++ scheduler/scheduler.go
@@ -83,12 +83,13 @@ func (e *Scheduler) EventLoop() {
 			e.Framework.FrameworkInfo.ID = event.Subscribed.GetFrameworkID()
 			e.Framework.MesosStreamID = res.Header.Get("Mesos-Stream-Id")
 
-			e.reconcile()
-			e.Redis.SaveFrameworkRedis(*e.Framework)
-			e.Redis.SaveConfig(*e.Config)
+			go e.reconcile()
+			go e.Redis.SaveFrameworkRedis(*e.Framework)
+			go e.Redis.SaveConfig(*e.Config)
 		case mesosproto.Event_UPDATE:
-			e.HandleUpdate(&event)
-			e.Redis.SaveConfig(*e.Config)
+			go e.HandleUpdate(&event)
+			go e.Redis.SaveConfig(*e.Config)
+			go e.callPluginEvent(event)
 		case mesosproto.Event_OFFERS:
 			// Search Failed containers and restart them
 			err = e.HandleOffers(event.Offers)
@@ -163,5 +164,25 @@ func (e *Scheduler) implicitReconcile() {
 
 	if err != nil {
 		logrus.WithField("func", "scheduler.implicitReconcile").Debug("Reconcile Error: ", err)
+	}
+}
+
+func (e *Scheduler) callPluginEvent(event mesosproto.Event) {
+	if e.Config.PluginsEnable {
+		for _, p := range e.Config.Plugins {
+			symbol, err := p.Lookup("Event")
+			if err != nil {
+				logrus.WithField("func", "scheduler.callPluginEvent").Error("Error lookup event function in plugin: ", err.Error())
+				continue
+			}
+
+			eventPluginFunc, ok := symbol.(func(mesosproto.Event))
+			if !ok {
+				logrus.WithField("func", "main.initPlugins").Error("Error plugin does not have init function")
+				continue
+			}
+
+			eventPluginFunc(event)
+		}
 	}
 }