+++ changelog.md
@@ -7,6 +7,8 @@
 - DEL: unneeded mesoscni env parameter.
 - ADD: Mesos Healthcheck integration. 
 - CHANGE: optimise offer handling
+- ADD: Posibility to add plugins. For an example, take a look into the plugins directory.
+- ADD: Kafka Plugin to forware mesos event messages to kafka
 
 ## 1.0.1